+++ Furball.Engine/Engine/Graphics/Drawables/Tweens/TweenTypes/DoubleTween.cs
@@ -22,7 +22,7 @@ namespace Furball.Engine.Engine.Graphics.Drawables.Tweens.TweenTypes {
         /// <param name="startTime">Start Time</param>
         /// <param name="endTime">End Time</param>
         /// <param name="easing">Easing</param>
-        public DoubleTween(TweenType type, double source, double dest, int startTime, int endTime, Easing easing = Easing.None) {
+        public DoubleTween(TweenType type, double source, double dest, double startTime, double endTime, Easing easing = Easing.None) {
             this.TweenType = type;
 
             this._startDouble = source;
